Actually test and fix the handling of unused `AsyncWriteSignal` (#15)
Unused `AsyncWriteSignal` with usage of unconstrained `await` on the
accompanied `ArcResource` will in fact hang SSR indefinitely as `Notify`
will never return. Provide an `impl Drop` for `feature = "ssr"` that
will notify the ready state so the detection of drop [as
documented](https://docs.rs/leptos_async_signal/0.4.0/leptos_async_signal/fn.async_signal.html)
will correctly function to provide the default value as intended, rather
than waiting forever. The `unused` test will validate this, albeit with
a timeout, so removing the `impl Drop` will result in the failure in
termination rather than waiting forever.

Also modified the `AsyncWriteSIgnal` to be an inner type enclosed in an
`Arc` and have the new outer type be the public type, so that
`Arc::get_mut` can be leveraged to ensure drop only happens on the final
copy, otherwise _any_ drop will invoke `.mark_ready()`, most certainly
prematurely, as per the `clone_drop` test. It's not possible to leverage
`AsyncState.inner` for this task because the circular dependency due to
the resource having copy of this and thus it will hold onto a copy of
the very thing it must somehow give up to continue (which it can't).

Alternatively, using timeout in the resource on `state.wait()` inside
the resource can also work, as it would best ensure it doesn't wait
forever, given how it may be possible to purposefully hide clones to not
letting the final drop to prevent this automatic ready state being set,
but that's kind of on the user.